+++ cpovc_ovc/functions.py
@@ -639,7 +639,7 @@ def save_viral_load(request):
 
 
 def method_once(method):
-    "A decorator that runs a method only once."
+    # A decorator that runs a method only once.
     attrname = "_%s_once_result" % id(method)
     def decorated(self, *args, **kwargs):
         try:
@@ -648,233 +648,4 @@ def method_once(method):
             setattr(self, attrname, method(self, *args, **kwargs))
             return getattr(self, attrname)
     return decorated
-
-
-class UpdateViralLoad(object):
-	"""docstring for UpdateViralLoad"""
-	def __init__(self):
-		self.api_url_base = settings.NASCOP_API_BASE_URL
-		self.login_url = settings.NASCOP_LOGIN_URL
-		self.email = settings.NASCOP_EMAIL
-		self.password = settings.NASCOP_PASSWORD
-		self.numbers_to_facilities_list = self.query_ccc_numbers_to_facilities()
-        # divide the data into 5 chunks.
-        self.five_chunks = self.chunk_it(numbers_to_facilities_list, 5)
-
-	
-	def query_ccc_numbers_to_facilities(self):
-		cursor = connection.cursor()
-		cursor.execute("SELECT ccc_number, facility_code, ovc_viral_load.person_id FROM eid, ovc_viral_load WHERE eid.person_id = ovc_viral_load.person_id AND ovc_viral_load.viral_load IS NULL;")
-		ccc_numbers_to_facilities = cursor.fetchall()
-		return ccc_numbers_to_facilities
-        
-    def chunk_it(self, seq, num):
-        # divide ccc_numbers-to-facilities list into manageable chunks
-    	avg = len(seq) / float(num)
-    	out = []
-    	last = 0.0
-
-    	while last < len(seq):
-    		out.append(seq[int(last):int(last + avg)])
-    		last += avg
-
-    	return out
-
-    def generate_token(self):
-        headers = {'Content-Type': 'application/x-www-form-urlencoded'}
-        credentials = { "email": self.email, "password": self.password }
-        response = requests.post(self.login_url, headers=headers, data=credentials)
-
-		if response.status_code == 200:
-			json_token = json.loads(response.content)
-			print(json.loads(response.content.decode('utf-8')))
-			api_token = json_token.get('token')
-			return api_token
-		else:
-			print(response)
-
-
-	def get_viral_load(self, token, facility, patientID):
-		headers = {'Authorization': 'Bearer {0}'.format(token)}
-		api_url = '{0}{1}/{2}'.format(self.api_url_base, facility, patientID)
-		response = requests.get(api_url, headers=headers)
-
-		if response.status_code == 200:
-			json_object = json.loads(response.content)
-			data_list = []
-			for test in json_object:
-				patient = test.get('PatientID')
-				date_tested = test.get('DateTested')
-				result = test.get('Result')
-				data = (patient, date_tested, result)
-				print("The details are: {0}, {1}, {2}".format(patient, date_tested, result))
-				data_list.append(data)
-			print(data_list)
-			return data_list
-		elif response.status_code == 404:
-			print('[!] [{0}] URL not found: [{1}]'.format(response.status_code,api_url))
-			return None
-		elif response.status_code == 401:
-			print('[!] [{0}] Authentication Failed'.format(response.status_code), response.content)
-			return None
-		elif response.status_code == 400:
-			print('[!] [{0}] Bad Request'.format(response.status_code), response.content)
-			return None
-		elif response.status_code >= 300:
-			print('[!] [{0}] Unexpected Redirect'.format(response.status_code), response.content)
-			return None
-		elif response.status_code >= 500:
-			print('[!] [{0}] Server Error'.format(response.status_code), response.content)
-			return None
-		else:
-			print('[?] Unexpected Error: [HTTP {0}]: Content: {1}'.format(response.status_code, response.content))
-			return None
-
-    def vl_to_int(self,string):
-		if "< LDL" in string:
-			return int(0)
-		elif "nvalid" in string:
-			return int(-998)
-		elif "ailed" in string:
-			return int(-999)
-		elif "ollect" in string:
-			return int(-997)
-		else:
-			return ''.join([i for i in string if i.isdigit()])
-
-	def loop_through_data(self, chunks):
-		try:
-            for chunk in chunks:
-                api_token = self.generate_token()
-                
-                for row in chunk:
-                    facility_code = row[1]
-                    ccc_number = row[0]
-                    person_id = row[2]
-                    facility_code_ccc = (facility_code, ccc_number)
-                    print(facility_code_ccc)
-                    viral_load_records = self.get_viral_load(api_token, facility_code, ccc_number)
-                    if viral_load_records:
-                        for record in viral_load_records:
-                            date_tested = record[1]
-                            result = record[2]
-                            result_to_int = self.vl_to_int(result)
-                            
-                            if OVCViralload.objects.filter(person_id=person_id, viral_date=viral_date, viral_load__isnull=True).exists():
-                                print("Entry contained in queryset")
-                                OVCViralload.objects.filter(person_id=person_id, viral_date=viral_date, viral_load__isnull=True).update(viral_load=result_to_int)
-                            else:
-                                vl = OVCViralload(person_id=person_id, viral_load=result_to_int, date=date_tested)
-                                vl.save()
-
-		except Exception as e:
-			print 'error exiting - %s' % (str(e))
-			raise e
-		else:
-			pass
-
-	
-
-class KMHFLFacilities(object):
-    '''
-        Auto-update the list of facilities from KMHFL
-    '''
-
-    def __init__(self):
-        self.username = settings.KMHFL_USERNAME
-        self.password = settings.KMHFL_PASSWORD
-        self.scope = settings.KMHFL_SCOPE
-        self.grant_type = settings.KMHFL_GRANT_TYPE
-        self.client_id = settings.KMHFL_CLIENTID
-        self.client_secret = settings.KMHFL_CLIENT_SECRET
-        self.api_base_url = settings.KMHFL_API_BASE_URL
-        self.facility_base_url = settings.KMHFL_FACILITY_BASE_URL
-        self.login_url = settings.KMHFL_LOGIN_URL
-        self.api_token = self.generate_token()
-        self.latest_facility = self.latest_facility()
-
-    @method_once
-    def latest_facility(self):
-        # query latest facility from db
-        latest_mfl_code = OVCFacility.objects.values("facility_code").exclude(facility_code__regex=r'[^0-9]').order_by('facility_code').last()
-        return latest_mfl_code["facility_code"]
-
-    @method_once
-    def generate_token(self):
-        # generate token.
-        login_url = self.login_url
-        headers = {'Content-Type': 'application/x-www-form-urlencoded'}
-        auth = (self.client_id, self.client_secret)
-        credentials = { "grant_type": self.grant_type, 
-                        "username": self.username, 
-                        "password": self.password, 
-                        "scope": self.scope }
-        response = requests.post(login_url, headers=headers, data=credentials, auth=auth)
-        if response.status_code == 200:
-            json_token = json.loads(response.content)
-            api_token = json_token.get('access_token')
-            return api_token
-        else:
-            print(response.content)
-    
-
-    def get_facilities(self):
-        # request for facilities
-        headers = {'Authorization': 'Bearer {0}'.format(self.api_token)}
-        api_url = '{0}facilities/facilities/?format=json'.format(self.api_base_url)
-        response = requests.get(api_url, headers=headers)
-        
-        if response.status_code == 200:
-            json_object = json.loads(response.content)
-            return json_object
-        else:
-            print(response.content, api_url)
-
-
-    def get_subcounty_id(self, facility_id):
-        # request facility and get sub-county id.
-        headers = {'Authorization': 'Bearer {0}'.format(self.api_token)}
-        payload = {'format': 'json'}
-        api_url = '{0}{1}'.format(self.facility_base_url, facility_id)
-        response = requests.get(api_url, headers=headers, params=payload)
-        if response.status_code == 200:
-            json_object = json.loads(response.content)
-            cpims_subcounty_id = json_object["constituency_code"]
-            return cpims_subcounty_id
-        else:
-            print(response, api_url)
-
-    @transaction.atomic
-    def get_newest_facilities(self):
-        # loop for new facilities.
-        try:
-            data = self.get_facilities()
-            results = data["results"]
-            for facility in results:
-                facility_code = facility["code"]
-                if facility_code > self.latest_facility:
-                    facility_id = facility["id"]
-                    cpims_subcounty_id = self.get_subcounty_id(facility_id)
-                    facility_name = facility["official_name"]
-                    sub_county_id = facility["sub_county_id"]
-                    new_facility = OVCFacility(facility_code=facility_code, 
-                        facility_name=facility_name, 
-                        sub_county_id=cpims_subcounty_id)
-                    
-                    new_facility.save()
-        
-        except Exception as e:
-            raise e
-        else:
-            pass
-
-    def schedule_update(self):
-        # Update facilities every tuesday at 00:15 am
-        schedule.every().tuesday.at("00:15").do(self.get_newest_facilities)
-
-        # # Check for pending schedules
-        # while True:
-        #     schedule.run_pending()
-        #     time.sleep(1)
-
